CIS4004Project-main/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
dj-database-url==2.1.0
mssql-django==1.3
pyodbc==4.0.39
python-dotenv==1.0.0
whitenoise==6.6.0



